+++ src/VegaXmlPlotter.cpp
@@ -14,14 +14,7 @@
 #include "TColor.h"
 #include "TTree.h"
 
-// #include "TBufferJSON.h"
-
 #include <thread>
-
-// #define VEGADEBUG 1
-
-
-
 
 void VegaXmlPlotter::init(){
 	DSCOPE();
@@ -36,29 +29,42 @@ void VegaXmlPlotter::init(){
 
 	setDefaultPalette();
 
-	handle_map[ "TCanvas" ] = &VegaXmlPlotter::exec_TCanvas;
-	handle_map[ "Data" ] 	= &VegaXmlPlotter::exec_Data;
-	handle_map[ "TFile" ] 	= &VegaXmlPlotter::exec_TFile;
-	handle_map[ "Plot" ] 	= &VegaXmlPlotter::exec_Plot;
-	handle_map[ "Loop" ] 	= &VegaXmlPlotter::exec_Loop;
-	handle_map[ "Axes" ] 	= &VegaXmlPlotter::exec_Axes;
-	handle_map[ "Export" ] 	= &VegaXmlPlotter::exec_Export;
+	handle_map[ "TCanvas"      ] = &VegaXmlPlotter::exec_TCanvas;
+	handle_map[ "Data"         ] = &VegaXmlPlotter::exec_Data;
+	handle_map[ "TFile"        ] = &VegaXmlPlotter::exec_TFile;
+	handle_map[ "Plot"         ] = &VegaXmlPlotter::exec_Plot;
+	handle_map[ "Loop"         ] = &VegaXmlPlotter::exec_Loop;
+	handle_map[ "Axes"         ] = &VegaXmlPlotter::exec_Axes;
+	handle_map[ "Export"       ] = &VegaXmlPlotter::exec_Export;
 	handle_map[ "ExportConfig" ] = &VegaXmlPlotter::exec_ExportConfig;
-	handle_map[ "Histo" ] 	= &VegaXmlPlotter::exec_Histo;
-	handle_map[ "TLine" ] 	= &VegaXmlPlotter::exec_TLine;
-	handle_map[ "TLatex" ] 	= &VegaXmlPlotter::exec_TLatex;
-	handle_map[ "TLegend" ] = &VegaXmlPlotter::exec_TLegend;
-	handle_map[ "Legend" ]  = &VegaXmlPlotter::exec_TLegend;
+	handle_map[ "Histo"        ] = &VegaXmlPlotter::exec_Histo;
+	handle_map[ "TLine"        ] = &VegaXmlPlotter::exec_TLine;
+	handle_map[ "TLatex"       ] = &VegaXmlPlotter::exec_TLatex;
+	handle_map[ "TLegend"      ] = &VegaXmlPlotter::exec_TLegend;
+	handle_map[ "Legend"       ] = &VegaXmlPlotter::exec_TLegend;
 
-	handle_map[ "Canvas" ]  = &VegaXmlPlotter::exec_Canvas;
-	handle_map[ "Pad" ]  = &VegaXmlPlotter::exec_Pad;
-	handle_map[ "Margins" ]  = &VegaXmlPlotter::exec_Margins;
-	// handle_map.insert( std::make_pair("TCanvas", &VegaXmlPlotter::exec_TCanvas) );
+	handle_map[ "Canvas"       ] = &VegaXmlPlotter::exec_Canvas;
+	handle_map[ "Pad"          ] = &VegaXmlPlotter::exec_Pad;
+	handle_map[ "Margins"      ] = &VegaXmlPlotter::exec_Margins;
 
+	handle_map[ "Transforms"   ] = &VegaXmlPlotter::exec_Transforms;
 
-}
-
-
+	handle_map[ "Projection"   ] = &VegaXmlPlotter::exec_transform_Projection;
+	handle_map[ "ProjectionX"  ] = &VegaXmlPlotter::exec_transform_ProjectionX;
+	handle_map[ "ProjectionY"  ] = &VegaXmlPlotter::exec_transform_ProjectionY;
+	handle_map[ "MultiAdd"     ] = &VegaXmlPlotter::exec_transform_MultiAdd;
+	handle_map[ "Add"          ] = &VegaXmlPlotter::exec_transform_Add;
+	handle_map[ "Divide"       ] = &VegaXmlPlotter::exec_transform_Divide;
+	handle_map[ "Rebin"        ] = &VegaXmlPlotter::exec_transform_Rebin;
+	handle_map[ "Scale"        ] = &VegaXmlPlotter::exec_transform_Scale;
+	handle_map[ "Draw"         ] = &VegaXmlPlotter::exec_transform_Draw;
+	handle_map[ "Clone"        ] = &VegaXmlPlotter::exec_transform_Clone;
+	handle_map[ "Smooth"       ] = &VegaXmlPlotter::exec_transform_Smooth;
+	handle_map[ "CDF"          ] = &VegaXmlPlotter::exec_transform_CDF;
+	handle_map[ "Style"        ] = &VegaXmlPlotter::exec_transform_Style;
+	handle_map[ "SetBinError"  ] = &VegaXmlPlotter::exec_transform_SetBinError;
+	handle_map[ "BinLabels"    ] = &VegaXmlPlotter::exec_transform_BinLabels;
+} // init
 
 void VegaXmlPlotter::exec_node( string _path ){
 	DSCOPE();
@@ -76,573 +82,6 @@ void VegaXmlPlotter::exec_node( string _path ){
 	exec( tag, _path );
 } // exec_node
 
-void VegaXmlPlotter::exec_TCanvas( string _path ){
-	DSCOPE();
-	int width = -1, height = -1;
-	
-	width = config.getInt( _path + ".TCanvas:width", width );
-	height = config.getInt( _path + ".TCanvas:height", height );
-
-	TCanvas * c = nullptr; 
-	if ( width != -1 && height > -1 ){
-		c = new TCanvas( "c", "c", width, height );
-		LOG_F( INFO, "TCanvas( width=%d, height=%d )", width, height );
-	}
-	else {
-		c = new TCanvas( "c" );
-		LOG_F( INFO, "TCanvas()" );
-	}
-
-	c->SetFillColor(0); 
-	c->SetFillStyle(4000);
-
-	c->cd();
-	c->Modified();
-	c->Draw(); 
-	c->Show();
-	//return c;
-} // exec_TCanvas
-
-void VegaXmlPlotter::exec_Data( string _path ){
-	DSCOPE();
-	if ( config.exists( _path + ":name" ) && !config.exists( _path + ":treeName" ) ){
-		loadDataFile( _path );
-	} else if ( config.exists( _path + ":name" ) && config.exists( _path + ":treeName" )){
-		loadChain( _path );
-	}
-} // exec_Data
-
-void VegaXmlPlotter::exec_TFile( string _path ){
-	DSCOPE();
-	string url = config.getString( _path + ":url" );
-	LOG_F( INFO, "Opening %s in RECREATE mode", url.c_str() );
-	dataOut = new TFile( url.c_str(), "RECREATE" );
-}
-
-void VegaXmlPlotter::exec_children( string _path ){
-	DSCOPE();
-	vector<string> paths = config.childrenOf( _path, 1 );
-	for ( string p : paths ){
-		exec_node( p );
-	}
-} //exec_children
-
-void VegaXmlPlotter::exec_children( string _path, string tag_type ){
-	DSCOPE();
-	vector<string> paths = config.childrenOf( _path, 1 );
-	DLOG( "exec children of %s, where tag = %s", _path.c_str(), tag_type.c_str() );
-	for ( string p : paths ){
-		string tag = config.tagName( p );
-
-		if ( tag_type == tag )
-			exec_node( p );
-	}
-} //exec_children
-
-void VegaXmlPlotter::exec_Loop( string _path ){
-	DSCOPE();
-	vector<string> states;
-	string var = "state";
-	string v = "%s";
-
-	if ( config.exists( _path + ":states" ) )
-		states = config.getStringVector( _path +":states" );
-
-	if ( config.exists( _path + ":var" ) )
-		var = config.getString( _path + ":var" );
-
-
-
-	for ( string state : states ){
-		DLOG( "Executing loop %s = %s", var.c_str(), state.c_str() );
-		string value = state;
-		config.set( var, value );
-		// vector<string> paths = config.childrenOf( _path, 1 );
-		exec_children( _path );
-	} // loop on states
-} // exec_Loop
-
-void VegaXmlPlotter::exec_Plot( string _path ) {
-	DSCOPE();
-	DLOG( "Makign Plot[%s] @ %s", config.getString( _path+":name", "" ).c_str(), _path.c_str() );
-
-	histos.clear();
-	graphs.clear();
-	funcs.clear();
-
-	exec_node( _path + ".Axes" );
-
-	// if ( config.exists( _path + ".Palette" ) ){
-	// 	gStyle->SetPalette( config.getInt( _path + ".Palette" ) );
-	// }
-
-	vector<string> tlp = { "Loop", "Histo", "Graph", "TF1", "TLine", "TLatex" };
-	vector<string> paths = config.childrenOf( _path, 1 );
-	for ( string p : paths ){
-		string tag = config.tagName( p );
-		if ( std::find( tlp.begin(), tlp.end(), tag ) != tlp.end() )
-			exec_node( p );
-	}
-
-	// Make LAtex on global scope for ease
-	exec_children( "", "TLatex" );
-
-
-
-	// makeLegend( _path, histos, graphs, funcs );
-	
-	exec_children( _path, "Legend" );
-	exec_children( _path, "Export" );
-
-	// cleanup this breaks Pads and doesnt really do anything useful, so taken out for now
-	// if ( nullptr != current_frame ){
-	// 	delete current_frame;
-	// 	current_frame = nullptr;
-	// }
-} // exec_Plot
-
-void VegaXmlPlotter::exec_Axes( string _path ){
-	DSCOPE();
-	
-	if ( !config.exists( _path ) ) return;
-
-	RooPlotLib rpl;
-	HistoBins x( config, _path );
-	HistoBins y( config, _path );
-
-	// also check for linspaces on x, y
-	LOG_F( 0, "Checking @%s:x :: %s ", _path.c_str(), config.oneOf( _path + ":x", _path + ":lsx" ).c_str() );
-	x.linspace( config, config.oneOf( _path + ":x", _path + ":lsx" ) );
-	x.arange( config, config.oneOf( _path + ":xrange", _path + ":xr" ) );
-
-	LOG_F( 0, "Checking @%s:y :: %s ", _path.c_str(), config.oneOf( _path + ":y", _path + ":lsy" ).c_str() );
-	y.linspace( config, config.oneOf( _path + ":y", _path + ":lsy" ) );
-	y.arange( config, config.oneOf( _path + ":yrange", _path + ":yr" ) );
-
-	// INFOC( "X : " << x.toString() << "\n\n\n" );
-	// INFOC( "Y : " << y.toString() );
-
-	if ( x.nBins() <= 0 ) {
-		LOG_F( ERROR, "X bins invalid, bailing out" );
-		// ERRORC( "Cannot make Axes, invalid x bins" );
-		return;
-	}
-
-	if ( y.nBins() <= 0 ) {
-		LOG_F( ERROR, "Y bins invalid, bailing out" );
-		// ERRORC( "Cannot make Axes, invalid y bins" );
-		return;
-	}
-
-	config.set( "x_min", dts(x.minimum()) );
-	config.set( "x_max", dts(x.maximum()) );
-
-	config.set( "y_min", dts(y.minimum()) );
-	config.set( "y_max", dts(y.maximum()) );
-
-	TH1 * frame = new TH1C( "frame", "", x.nBins(), x.bins.data() );
-
-	rpl.style( frame ).set( "yr", y.minimum(), y.maximum() );
-	rpl.set( config, _path ).set( config, _path + ":style" ).set( config, _path + ".style" );
-
-	frame->SetMarkerStyle(8);
-	frame->SetMarkerSize(0);
-	frame->Draw("p");
-
-	current_frame = (TH1C*)frame;
-} // exec_Axes
-
-void VegaXmlPlotter::exec_Export( string _path ){
-	DSCOPE();
-	TPad * _pad = (TPad*)gPad;
-	
-	if ( !config.exists( _path + ":url" ) ) return;
-
-	string url = config.getXString( _path + ":url" );
-	if ( url.find( ".json" ) != string::npos ){
-		// TBufferJSON::ExportToFile( url.c_str(), _pad );
-		// LOG_F( INFO, "%s",  );
-		// ofstream fout( url.c_str() );
-		// fout << TBufferJSON::ConvertToJSON( _pad ).Data();
-		// fout.close();
-	} else {
-		_pad->Print( url.c_str() );
-	}
-} // exec_Export
-
-void VegaXmlPlotter::exec_Histo( string _path ){
-	DSCOPE();
-	RooPlotLib rpl;
-
-	TH1* h = findHistogram( _path, 0 );
-	LOG_F( INFO, "Found Histogram @%s = %p", _path.c_str(), h );
-
-	if ( nullptr == h ) return;
-
-	config.set( "ClassName", h->ClassName() );
-	DLOG( "ClassName %s", config[ "ClassName" ].c_str() );
-	
-	string name = config.getXString( _path + ":name" );
-	string data = config.getXString( _path + ":data" );
-	
-	string fqn = fullyQualifiedName( data, name );
-
-	float nI = h->Integral();
-	string normp = config.oneOf( _path + ".Norm", _path + ":norm" );
-	if ( "" != normp && nI > 0 && config.getBool( normp, true ) == true){
-		float normC = config.get<float>( normp, 0 );
-		if ( normC <= 0 )
-			normC = 1.0;
-		LOG_F( INFO, "Normalize: %s to %f / %f", h->GetName(), normC, nI );
-		h->Scale( normC / nI );
-	}
-
-	string styleRef = config.getXString( _path + ":style" );
-	
-	if ( config.exists( styleRef ) ){
-		rpl.style( h ).set( config, styleRef );
-	}
-
-	rpl.style( h ).set( config, _path ).set( config, _path + ".style" ).draw();
-
-	if ( config.exists( _path +":after_draw" ) ){
-		string cmd = ".x " + config[_path+":after_draw"] + "( " + h->GetName() + " )";
-		LOG_F( INFO, "Executing: %s", cmd.c_str()  );
-		gROOT->ProcessLine( cmd.c_str() );
-	}
-
-	TPaveStats *st = (TPaveStats*)h->FindObject("stats");
-	if ( nullptr != st  ){
-		positionOptStats( _path, st );
-		// st->SetX1NDC( 0.7 ); st->SetX2NDC( 0.975 );
-	}
-
-
-	histos[ nameOnly(fqn) ] = h;
-	histos[ fqn ] 			= h;
-
-	// return h;
-} // exec_Histo
-
-void VegaXmlPlotter::exec_TLine( string _path){
-	DSCOPE();
-
-	vector<double> x(2);
-	vector<double> y(2);
-
-	if ( config.exists( _path + ":x" ) ){
-		x = config.getDoubleVector( _path + ":x" );
-	}
-	if ( config.exists( _path + ":y" ) ){
-		y = config.getDoubleVector( _path + ":y" );
-	}
-	if ( config.exists( _path + ":p1" ) ){
-		vector<double> p = config.getDoubleVector( _path + ":p1" );
-		if ( p.size() >= 2 ){
-			x[0] = p[0];
-			y[0] = p[1];
-		}
-	}
-	if ( config.exists( _path + ":p2" ) ){
-		vector<double> p = config.getDoubleVector( _path + ":p2" );
-		if ( p.size() >= 2 ){
-			x[1] = p[0];
-			y[1] = p[1];
-		}
-	}
-
-	LOG_F( INFO, "Line (%f, %f)->(%f, %f)", x[0], y[0], x[1], y[1] );
-	TLine * line = new TLine( 
-		config.getDouble( _path + ":x1", x[0] ),
-		config.getDouble( _path + ":y1", y[0] ),
-		config.getDouble( _path + ":x2", x[1] ),
-		config.getDouble( _path + ":y2", y[1] ) );
-
-	line->SetLineColor( color( config.getString( _path + ":color" ) ) );
-	line->SetLineWidth( config.getInt( _path + ":width", 1 ) );
-	line->SetLineStyle( config.getInt( _path + ":style", 2 ) );
-
-	line->Draw("same");
-} // exec_TLine
-
-void VegaXmlPlotter::exec_TLatex( string _path ){
-	DSCOPE();
-
-	TLatex latex;
-	if ( !config.exists( _path + ":text" ) ) return;
-
-	string text = config.getXString( _path + ":text" );
-	
-	latex.SetTextSize( config.getFloat( _path + ":size", 0.05 ) );
-	if ( config.exists( _path + ":point" ) )
-		latex.SetTextSize( config.getFloat( _path + ":point", 18 ) / 360.0 );
-	
-
-
-	if ( config.exists( _path +":color" ) ){
-		string cs = config.getXString( _path + ":color" );
-		latex.SetTextColor( color( cs ) );
-	} else {
-		latex.SetTextColor( kBlack );
-	}
-	
-	TLatex * TL = nullptr;
-	if ( config.exists( _path + ":x" ) && config.exists( _path + ":y" ) ){
-		TL = latex.DrawLatexNDC( config.getFloat( _path + ":x" ), 
-						config.getFloat( _path + ":y" ), 
-						config.getXString( _path + ":text" ).c_str() );
-	} else if ( config.exists( _path + ":ux" ) && config.exists( _path + ":uy" ) ){
-		TL = latex.DrawLatex( config.getFloat( _path + ":ux" ), 
-						config.getFloat( _path + ":uy" ), 
-						config.getXString( _path + ":text" ).c_str() );
-	}
-	if ( nullptr == TL ){
-		LOG_F(ERROR, "Cannot draw latex, missing attribute. Must have pad corrds (:x,:y) or user corrds( :ux, :uy)" ); 
-		return;
-	}		 
-
-	if ( config.exists( _path + ":font" ) ){
-		TL->SetTextFont( config.getInt( _path + ":font" ) );
-	}
-	if ( config.exists( _path + ":angle" ) ){
-		TL->SetTextAngle( config.getFloat( _path + ":angle" ) );
-	}
-
-	if ( config.exists( _path + ":align" ) ){
-		string strAlign = config.getString( _path + ":align" );
-		if ( "top" == strAlign || "t" == strAlign ){
-			TL->SetTextAlign( 13 );
-		} else if ( "center" == strAlign || "centered" == strAlign || "c" == strAlign ){
-			TL->SetTextAlign( 12 );
-		} else if ( "bottom" == strAlign || "b" == strAlign ){
-			TL->SetTextAlign( 11 );
-		} else if ( "special" == strAlign || "s" == strAlign ){
-			TL->SetTextAlign( 10 );
-		} else {
-			TL->SetTextAlign( config.getInt( _path + ":align" ) );
-		}
-	}
-} // exec_TLatex
-
-void VegaXmlPlotter::exec_TLegend( string _path ){
-	DSCOPE();
-	
-	RooPlotLib rpl;
-	float x1, y1, x2, y2;
-	
-	x1 = config.getDouble( _path + ".Position:x1", 0.1 );
-	y1 = config.getDouble( _path + ".Position:y1", 0.7 );
-	x2 = config.getDouble( _path + ".Position:x2", 0.5 );
-	y2 = config.getDouble( _path + ".Position:y2", 0.9 );
-
-	string spos = config.getXString( _path + ".Position:pos" );
-
-	float w = config.getFloat( _path + ".Position:w", 0.4 );
-	float h = config.getFloat( _path + ".Position:h", 0.2 );
-	vector<float> padding = config.getFloatVector( _path + ".Position:padding" );
-	if ( padding.size() < 4 ){
-		padding.clear();
-		padding.push_back( 0 ); padding.push_back( 0 );
-		padding.push_back( 0 ); padding.push_back( 0 );
-	}
-
-	DLOG( "Legend padding : %s", vts( padding ).c_str() );
-
-
-	if ( spos.find("top") != string::npos ){
-		DLOG( "LEGEND: TOP" );
-		y2 = 0.9 - padding[0];
-		if ( nullptr != gPad) y2 = 1.0 - gPad->GetTopMargin() - padding[0];
-		y1 = y2 - h;
-	}
-	if ( spos.find("right") != string::npos ){
-		DLOG( "LEGEND: RIGHT" );
-		x2 = 0.9 - padding[1];
-		if ( nullptr != gPad) x2 = 1.0 - gPad->GetRightMargin() - padding[1];
-		x1 = x2 - w;
-	}
-	if ( spos.find("bottom") != string::npos ){
-		DLOG( "LEGEND: BOTTOM" );
-		y1 = 0.1 + padding[2];
-		if ( nullptr != gPad) y1 = gPad->GetBottomMargin() + padding[2];
-		y2 = y1 + h;
-	}
-	if ( spos.find("left") != string::npos ){
-		DLOG( "LEGEND: LEFT" );
-		x1 = 0.1 + padding[3];
-		if ( nullptr != gPad) x1 = gPad->GetLeftMargin() + padding[3];
-		x2 = x1 + w;
-	}
-	if ( spos.find("vcenter") != string::npos ){
-		DLOG( "LEGEND: VERTICAL CENTER" );
-		y1 = 0.5 - h/2;
-		y2 = 0.5 + h/2;
-	}
-	if ( spos.find("hcenter") != string::npos ){
-		DLOG( "LEGEND: HORIZONTAL CENTER" );
-		x1 = 0.5 - w/2;
-		x2 = 0.5 + w/2;
-	}
-	DLOG( "Legend position: (%f, %f) -> (%f, %f)", x1, y1, x2, y2 );
-	TLegend * leg = new TLegend( x1, y1, x2, y2 );
-
-	if ( config.exists(  _path + ":title" ) ) 
-		leg->SetHeader( config.getXString( _path + ":title" ).c_str() );
-	leg->SetNColumns( config.getInt( _path + ":columns", 1 ) );
-
-	vector<string> entries = config.childrenOf( _path + "", "Entry" );
-
-	// HISTOGRAMS
-	for ( string entryPath : entries ){
-		//INFO( classname(), "Entry @" << entryPath );
-		if ( config.exists( entryPath + ":name" ) != true ) continue;
-		string name = config.getXString( entryPath + ":name" );
-		
-		if ( histos.count( name ) <= 0 || histos[ name ] == nullptr ) {
-			continue;
-		}
-		TH1 * h = (TH1*)histos[ name ]->Clone( ("hist_legend_" + name).c_str() );
-		
-		string t = config.getXString(  entryPath + ":title", name );
-		string opt = config.getXString(  entryPath + ":opt", "l" );
-		rpl.style( h ).set( config, entryPath );
-		
-
-
-		leg->AddEntry( h, t.c_str(), opt.c_str() );
-	}
-
-	// GRAPHS
-	for ( string entryPath : entries ){
-		if ( config.exists( entryPath + ":name" ) != true ) continue;
-		string name = config.getXString( entryPath + ":name" );
-		if ( graphs.count( name ) <= 0 || graphs[ name ] == nullptr )
-			continue;
-		TGraph * g = (TGraph*)graphs[ name ]->Clone( ("graph_legend_" + name).c_str() );
-
-		string t = config.getXString(  entryPath + ":title", name );
-		string opt = config.getXString(  entryPath + ":opt", "l" );
-		rpl.style( g ).set( config, entryPath );
-		leg->AddEntry( g, t.c_str(), opt.c_str() );
-	}
-
-	// FUNCTIONS
-	for ( string entryPath : entries ){
-		//INFO( classname(), "Entry @" << entryPath );
-		if ( config.exists( entryPath + ":name" ) != true ) continue;
-		string name = config.getXString( entryPath + ":name" );
-		
-		if ( funcs.count( name ) <= 0 || funcs[ name ] == nullptr ) {
-			continue;
-		}
-		TF1 * f = (TF1*)funcs[ name ]->Clone( ("graph_legend_" + name).c_str() );
-
-
-		//INFO( classname(), "Entry histo=" << h );
-		string t = config.getXString(  entryPath + ":title", name );
-		string opt = config.getXString(  entryPath + ":opt", "l" );
-
-		rpl.style( f ).set( config, entryPath );
-		leg->AddEntry( f, t.c_str(), opt.c_str() );
-	}
-
-	if ( entries.size() <= 0 ){
-		// Add an entry for all histograms by default if no <Entry> nodes found
-		for ( auto kv : histos ){
-			string t = config.getString(  _path + "." + kv.first+ ":title", kv.first) ;
-			string opt = config.getString(  _path + "." + kv.first+ ":opt", "l") ;
-			leg->AddEntry( kv.second, t.c_str(), opt.c_str() );
-		}
-	}
-
-	if ( config.exists( _path + ":text_size" ) ){
-		leg->SetTextSize( config.getDouble( _path + ":text_size" ) );
-	}
-	if ( config.exists( _path + ":text_point" ) ){
-		leg->SetTextSize( config.getDouble( _path + ":text_point" ) / 360.0 );
-	}
-	if ( config.exists( _path + ":border_size" ) ){
-		leg->SetBorderSize( config.getDouble( _path + ":border_size" ) );
-	}
-	if ( config.exists( _path + ":fill_color" ) ){
-		leg->SetFillColor( color( config.getString( _path + ":fill_color" ) ) );
-	}
-	if ( config.exists( _path + ":fill_style" ) ){
-		//INFOC( "FillStyle=" << config.getInt( _path + ":fill_style" ) );
-		leg->SetFillStyle( config.getInt( _path + ":fill_style" ) );
-	}
-
-	leg->Draw( );
-} // exec_TLegend
-
-void VegaXmlPlotter::exec_Canvas( string _path ){
-	DSCOPE();
-
-	xcanvas = new XmlCanvas( config, _path );
-
-	exec_children( _path, "Loop" );
-	exec_children( _path, "Pad" );
-	
-	xcanvas->getCanvas()->Modified();
-	xcanvas->getCanvas()->Update();
-
-	gPad = (TPad*) xcanvas->getCanvas();
-	exec_children( _path, "Export" );
-} // exec_Canvas
-
-void VegaXmlPlotter::exec_Pad( string _path ){
-	DSCOPE();
-	if ( nullptr == xcanvas ){
-		LOG_F( ERROR, "xcanvas is NULL, cannot make Pad @ %s", _path.c_str() );
-		return;
-	}
-	string n = config.getString( _path + ":name" );
-	if ( "" == n ){
-		LOG_F( WARNING, "Pad must have valid name, skipping pad @ %s", _path.c_str() );
-		return;
-	}
-
-	XmlPad * xpad = xcanvas->activatePad( n );
-
-	// move pad to origin
-	xpad->moveToOrigin();
-	xpad->getRootPad()->Update();
-	
-	exec_Plot( _path );
-
-	xpad->reposition();
-	xpad->getRootPad()->Update();
-}
-
-void VegaXmlPlotter::exec_Margins( string _path ){
-	DSCOPE();
-	
-	float tm=-1, rm=-1, bm=-1, lm=-1;
-
-	vector<float> parts = config.getFloatVector( _path );
-	if ( parts.size() >= 4 ){
-		tm=parts[0];rm=parts[1];bm=parts[2];lm=parts[3];
-	}
-
-	tm = config.getFloat( _path + ":top"    , tm );
-	rm = config.getFloat( _path + ":right"  , rm );
-	bm = config.getFloat( _path + ":bottom" , bm );
-	lm = config.getFloat( _path + ":left"   , lm );
-
-	LOG_F( INFO, "Margins( %f, %f, %f, %f )", tm, rm, bm, lm );
-
-	if ( tm>=0 ) gPad->SetTopMargin( tm );
-	if ( rm>=0 ) gPad->SetRightMargin( rm );
-	if ( bm>=0 ) gPad->SetBottomMargin( bm );
-	if ( lm>=0 ) gPad->SetLeftMargin( lm );
-} // exec_Margins
-
-void VegaXmlPlotter::exec_ExportConfig( string _path ){
-	DSCOPE();
-	LOG_F( INFO, "exporting config to %s", config[ _path + ":url" ].c_str() );
-	config.deleteNode( _path );	// not working!!!
-	config.toXmlFile( config[_path+":url"] );
-} // ExportConfig
 
 void VegaXmlPlotter::make(){
 	DSCOPE();
@@ -678,54 +117,6 @@ void VegaXmlPlotter::make(){
 		LOG_F( INFO, "Write to %s completed", config.getString( "TFile:url" ).c_str() );
 	}
 } // make
-
-
-
-
-
-
-// void VegaXmlPlotter::make() {
-// 	DSCOPE();
-
-// 	TCanvas *c = new TCanvas( "rp" );
-// 	// c->Print( "rp.pdf[" );
-
-// 	loadData();
-// 	makeOutputFile();
-
-// 	// string ll = config.getString( "Logger:globalLogLevel", "debug" );
-// 	//INFOC( "Setting ll to " << quote(ll) );
-// 	// Logger::setGlobalLogLevel( "none" );
-// 	makeTransforms();
-// 	makePlots();
-// 	makePlotTemplates();
-// 	makeCanvases();
-
-// 	if ( dataOut ){
-// 		LOG_F( INFO, "Writing data to %s", config.getXString( "TFile:url" ).c_str() );
-// 		dataOut->Write();
-// 		dataOut->Close();
-// 		LOG_F( INFO, "Write completed" );
-// 	}
-
-// 	// c->Print( "rp.pdf]" );
-
-// 	if ( config.exists( "ExportConfig:url" ) ){
-// 		LOG_F( INFO, "exporting config to %s", config["ExportConfig:url"].c_str() );
-// 		config.deleteNode( "ExportConfig" );	// not working!!!
-// 		config.toXmlFile( config["ExportConfig:url"] );
-// 	}
-	
-// 	// config.dumpToFile( "config_dump.txt" );
-// }
-
-void VegaXmlPlotter::makeHandlers(){
-
-	// vector<string> paths = config.childrenOf( nodePath );
-	// for ( string path : paths ){
-	// 	LOG_F( INFO, "%s -> %s", path.c_str(), config.tagName(path).c_str() );
-	// }
-}
 
 map<string, shared_ptr<TF1> > VegaXmlPlotter::makeTF(string _path){
 	DSCOPE();
@@ -806,13 +197,12 @@ void VegaXmlPlotter::inlineDataFile( string _path, TFile *_f ){
 	config.deleteAttribute( _path + ":url" );
 
 	// LOG_F( INFO, "resulting XML to inline : \n%s", xml.c_str() );
-}
+} // inlineDataFile
 
 int VegaXmlPlotter::numberOfData() {
 	DSCOPE();
 	return dataFiles.size() + dataChains.size();
-}
-
+} // numberOfData
 
 void VegaXmlPlotter::loadChain( string _path ){
 	DSCOPE();
@@ -838,7 +228,7 @@ void VegaXmlPlotter::loadChain( string _path ){
 	if ( dataChains[name]->GetListOfFiles() )
 		nFiles = dataChains[name]->GetListOfFiles()->GetEntries();
 	LOG_S(INFO) << "Chain has " << nFiles << plural( nFiles, " file", " files" );
-}
+} // loadChain
 
 void VegaXmlPlotter::loadData(){
 	DSCOPE();
@@ -856,140 +246,7 @@ void VegaXmlPlotter::loadData(){
 	if ( data_nodes.size() <= 0 ){
 		LOG_F(WARNING, "No data nodes found!" );
 	}
-
 } // loadData
-
-
-// void VegaXmlPlotter::makeOutputFile(){
-// 	DSCOPE();
-// 	if ( config.exists( "TFile:url" ) ){
-// 		dataOut = new TFile( config.getXString( "TFile:url" ).c_str(), "RECREATE" );
-// 	} else {
-// 		//INFO( classname(), "No output data file requested" );
-// 	}
-// }
-
-// void VegaXmlPlotter::makePlotTemplates(){
-// 	DSCOPE();
-// 	vector<string> plott_paths = config.childrenOf( nodePath, "PlotTemplate" );
-// 	DLOG_F( INFO, "Found %lu PlotTemplate", plott_paths.size() );
-// 	for ( string path : plott_paths ){
-// 		vector<string> names = config.getStringVector( path + ":names" );
-// 		if ( names.size() <= 1 ){
-// 			names = glob( config.getString( path+":names" ) );
-// 		}
-
-// 		TCanvas * c = makeCanvas( path );
-
-// 		int iPlot = 0;
-// 		for ( string name : names ){
-
-// 			// set the global vars for this template
-// 			config.set( "name", name );
-// 			config.set( "uname", underscape(name) );
-// 			config.set( "iPlot", ts(iPlot) );
-
-// 			makeMargins( path );
-// 			makePlot( path );
-// 			iPlot++;
-// 		}
-// 	}
-// }
-
-// void VegaXmlPlotter::makePlot( string _path, TPad * _pad ){
-// 	DSCOPE();
-// 	LOG_F( INFO, "Makign Plot[%s] @ %s", config.getString( _path+":name", "" ).c_str(), _path.c_str() );
-
-// 	 TODO
-// 	 * change logic so that it renders in the order of the nodes, 
-// 	 * not All Hists, all graphs, all functions -> this makes it impossible to controll whats on top of what
-	 
-
-// 	if ( config.exists( _path + ".Palette" ) ){
-// 		gStyle->SetPalette( config.getInt( _path + ".Palette" ) );
-// 	}
-
-// 	makeAxes( _path + ".Axes" );
-// 	map<string, TH1*> histos = makeHistograms( _path );
-// 	map<string, TGraph*> graphs = makeGraphs( _path );
-// 	map<string, shared_ptr<TF1>> funcs =  makeTF( _path );
-// 	// makeHistoStack( histos );
-// 	makeLatex(""); // global first
-// 	makeLatex(_path);
-// 	makeLine( _path );
-// 	makeLegend( _path, histos, graphs, funcs );
-
-// 	makeExports( _path, _pad );
-// 	// INFOC( "Finished making Plot" );
-// }
-
-// void VegaXmlPlotter::makePlots(){
-// 	DSCOPE();
-// 	vector<string> plot_paths = config.childrenOf( nodePath, "Plot" );
-	
-// 	DLOG_S(INFO) << "Found " << plot_paths.size() << plural( plot_paths.size(), " Plot", " Plots" );
-
-// 	for ( string path : plot_paths ){
-// 		TCanvas * c = makeCanvas( path ); 
-// 		c->cd();
-// 		makeMargins( path );
-// 		makePlot( path );
-// 	}
-// }
-
-// void VegaXmlPlotter::makeCanvases(){
-// 	DSCOPE();
-// 	vector<string> c_paths = config.childrenOf( nodePath, "Canvas" );
-// 	//INFO( classname(), "Found " << c_paths.size() << plural( c_paths.size(), " Canvas", " Canvases" ) );
-// 	for ( string path : c_paths ){
-// 		drawCanvas( path );
-// 	}
-// }
-
-// void VegaXmlPlotter::makeHistoStack( map<string, TH1*> histos ){
-// DSCOPE();
-// 	THStack *hs = new THStack( "hs", "" );
-// 	for ( auto kv : histos ){
-// 		hs->Add( kv.second );
-// 	}
-// 	RooPlotLib rpl;
-// 	rpl.style( hs )
-// }
-
-// void VegaXmlPlotter::makeMargins( string _path ){
-// 	DSCOPE();
-	
-// 	float tm=-1, rm=-1, bm=-1, lm=-1;
-
-// 	vector<float> parts = config.getFloatVector( "Margins" );
-// 	if ( parts.size() >= 4 ){
-// 		tm=parts[0];rm=parts[1];bm=parts[2];lm=parts[3];
-// 	}
-
-// 	parts = config.getFloatVector( _path + ".Margins" );
-// 	if ( parts.size() >= 4 ){
-// 		tm=parts[0];rm=parts[1];bm=parts[2];lm=parts[3];
-// 	}
-
-// 	// Global first
-	
-// 	tm = config.getFloat( "Margins:top"    , tm );
-// 	rm = config.getFloat( "Margins:right"  , rm );
-// 	bm = config.getFloat( "Margins:bottom" , bm );
-// 	lm = config.getFloat( "Margins:left"   , lm );
-
-// 	tm = config.getFloat( _path + ".Margins:top"    , tm );
-// 	rm = config.getFloat( _path + ".Margins:right"  , rm );
-// 	bm = config.getFloat( _path + ".Margins:bottom" , bm );
-// 	lm = config.getFloat( _path + ".Margins:left"   , lm );
-
-// 	LOG_F( INFO, "Margins( %f, %f, %f, %f )", tm, rm, bm, lm );
-// 	if ( tm>=0 ) gPad->SetTopMargin( tm );
-// 	if ( rm>=0 ) gPad->SetRightMargin( rm );
-// 	if ( bm>=0 ) gPad->SetBottomMargin( bm );
-// 	if ( lm>=0 ) gPad->SetLeftMargin( lm );
-// }
-
 
 TObject* VegaXmlPlotter::findObject( string _path ){
 	DSCOPE();
@@ -1025,9 +282,7 @@ TObject* VegaXmlPlotter::findObject( string _path ){
 
 	LOG_F( WARNING, "Object [%s/%s] was not found in any source", data.c_str(), name.c_str() );
 	return nullptr;
-}
-
-
+} // findObject
 
 TH1 *VegaXmlPlotter::findHistogram( string _data, string _name ){
 	DSCOPE();
@@ -1056,7 +311,7 @@ TH1 *VegaXmlPlotter::findHistogram( string _data, string _name ){
 	}
 
 	return nullptr;
-}
+} // findHistogram
 
 TH1* VegaXmlPlotter::findHistogram( string _path, int iHist, string _mod ){
 	DSCOPE();
@@ -1107,148 +362,7 @@ TH1* VegaXmlPlotter::findHistogram( string _path, int iHist, string _mod ){
 
 
 	return nullptr;
-}
-
-
-TH1* VegaXmlPlotter::makeAxes( string _path ){
-	DSCOPE();
-	
-	if ( !config.exists( _path ) ) return nullptr;
-
-	RooPlotLib rpl;
-	HistoBins x( config, _path );
-	HistoBins y( config, _path );
-
-	// also check for linspaces on x, y
-	LOG_F( 0, "Checking @%s:x :: %s ", _path.c_str(), config.oneOf( _path + ":x", _path + ":lsx" ).c_str() );
-	x.linspace( config, config.oneOf( _path + ":x", _path + ":lsx" ) );
-	x.arange( config, config.oneOf( _path + ":xrange", _path + ":xr" ) );
-
-	LOG_F( 0, "Checking @%s:y :: %s ", _path.c_str(), config.oneOf( _path + ":y", _path + ":lsy" ).c_str() );
-	y.linspace( config, config.oneOf( _path + ":y", _path + ":lsy" ) );
-	y.arange( config, config.oneOf( _path + ":yrange", _path + ":yr" ) );
-
-	// INFOC( "X : " << x.toString() << "\n\n\n" );
-	// INFOC( "Y : " << y.toString() );
-
-	if ( x.nBins() <= 0 ) {
-		LOG_F( ERROR, "X bins invalid, bailing out" );
-		// ERRORC( "Cannot make Axes, invalid x bins" );
-		return nullptr;
-	}
-
-	if ( y.nBins() <= 0 ) {
-		LOG_F( ERROR, "Y bins invalid, bailing out" );
-		// ERRORC( "Cannot make Axes, invalid y bins" );
-		return nullptr;
-	}
-
-	config.set( "x_min", dts(x.minimum()) );
-	config.set( "x_max", dts(x.maximum()) );
-
-	config.set( "y_min", dts(y.minimum()) );
-	config.set( "y_max", dts(y.maximum()) );
-
-	TH1 * frame = new TH1C( "frame", "", x.nBins(), x.bins.data() );
-
-	rpl.style( frame ).set( "yr", y.minimum(), y.maximum() );
-	rpl.set( config, _path ).set( config, _path + ":style" ).set( config, _path + ".style" );
-
-	frame->SetMarkerStyle(8);
-	frame->SetMarkerSize(0);
-	frame->DrawClone("p");
-
-	return frame;
-}
-
-
-// map<string, TH1*> VegaXmlPlotter::makeHistograms( string _path ){
-// 	DSCOPE();
-// 	map<string, TH1*> histos;
-// 	vector<string> hist_paths = config.childrenOf( _path, "Histo" );
-
-// 	LOG_F( INFO, "Found %lu Histograms", hist_paths.size() );
-// 	for ( string hpath : hist_paths ){
-
-// 		if ( config.exists( hpath + ":names" ) ){
-// 			//INFOC( "HISTOGRAM TEMPLATE at " << hpath );
-
-// 			vector<string> names = config.getStringVector( hpath + ":names" );
-// 			if ( names.size() <= 1 ){
-// 				names = glob( config.getString( hpath + ":names" ) );
-// 			}
-
-// 			int i = 0;
-// 			for ( string name : names ){
-
-// 				// set the global vars for this template
-// 				config.set( hpath + ":name", name );
-// 				config.set( "i", ts(i) );
-
-// 				//INFOC( "Histogram from template name = " << config.getXString( hpath + ":name" ) );
-
-// 				string fqn = "";
-// 				TH1 * h = makeHistogram( hpath, fqn );
-// 				histos[ nameOnly(fqn) ] = h;
-// 				histos[ fqn ] = h;
-// 				i++;
-// 			}
-// 			continue;
-// 		}
-
-// 		string fqn = "";
-// 		TH1 * h = makeHistogram( hpath, fqn );
-// 		histos[ nameOnly(fqn) ] = h;
-// 		histos[ fqn ] = h;
-// 		//INFOC( "[" << nameOnly(fqn) << "] = " << h );
-// 		//INFOC( "[" << fqn << "] = " << h );
-// 	}
-
-// 	return histos;
-// }
-
-
-map<string, TGraph*> VegaXmlPlotter::makeGraphs( string _path ){
-	DSCOPE();
-	map<string, TGraph*> graphs;
-	vector<string> paths = config.childrenOf( _path, "Graph" );
-
-	for ( string gpath : paths ){
-
-		if ( config.exists( gpath + ":names" ) ){
-			//INFOC( "HISTOGRAM TEMPLATE at " << gpath );
-
-			vector<string> names = config.getStringVector( gpath + ":names" );
-			if ( names.size() <= 1 ){
-				names = glob( config.getString( gpath + ":names" ) );
-			}
-
-			int i = 0;
-			for ( string name : names ){
-
-				// set the global vars for this template
-				config.set( gpath + ":name", name );
-				config.set( "i", ts(i) );
-
-				//INFOC( "Histogram from template name = " << config.getXString( gpath + ":name" ) );
-
-				string fqn = "";
-				TGraph * h = makeGraph( gpath, fqn );
-				graphs[ nameOnly(fqn) ] = h;
-				graphs[ fqn ] = h;
-				i++;
-			}
-			continue;
-		} else {
-			string fqn = "";
-			TGraph * h = makeGraph( gpath, fqn );
-			graphs[ nameOnly(fqn) ] = h;
-			graphs[ fqn ] = h;
-		}
-	}
-
-	return graphs;
-}
+} //findHistogram
 
 TGraph* VegaXmlPlotter::makeGraph( string _path, string &fqn ){
 	DSCOPE();
@@ -1276,7 +390,7 @@ TGraph* VegaXmlPlotter::makeGraph( string _path, string &fqn ){
 	rpl.style( g ).set( config, _path ).set( config, _path + ":style" ).set( config, _path + ".style" ).draw();
 
 	return g;
-}
+} // makeGraph
 
 TH1* VegaXmlPlotter::makeHistogram( string _path, string &fqn ){
 	DSCOPE();
@@ -1330,313 +444,7 @@ TH1* VegaXmlPlotter::makeHistogram( string _path, string &fqn ){
 		// st->SetX1NDC( 0.7 ); st->SetX2NDC( 0.975 );
 	}
 	return h;
-}
-
-
-void VegaXmlPlotter::makeLegend( string _path, map<string, TH1*> &histos, map<string, TGraph*> &graphs, map<string, shared_ptr<TF1>> &funcs ){
-	DSCOPE();
-	if ( config.exists( _path + ".Legend" ) ){
-		
-		RooPlotLib rpl;
-		float x1, y1, x2, y2;
-		x1 = config.getDouble( _path + ".Legend.Position:x1", 0.1 );
-		y1 = config.getDouble( _path + ".Legend.Position:y1", 0.7 );
-		x2 = config.getDouble( _path + ".Legend.Position:x2", 0.5 );
-		y2 = config.getDouble( _path + ".Legend.Position:y2", 0.9 );
-
-		string spos = config.getXString( _path + ".Legend.Position:pos" );
-
-		float w = config.getFloat( _path + ".Legend.Position:w", 0.4 );
-		float h = config.getFloat( _path + ".Legend.Position:h", 0.2 );
-		vector<float> padding = config.getFloatVector( _path + ".Legend.Position:padding" );
-		if ( padding.size() < 4 ){
-			padding.clear();
-			padding.push_back( 0 ); padding.push_back( 0 );
-			padding.push_back( 0 ); padding.push_back( 0 );
-		}
-
-		DLOG( "Legend padding : %s", vts( padding ).c_str() );
-
-
-		if ( spos.find("top") != string::npos ){
-			LOG_F( INFO, "LEGEND: TOP" );
-			y2 = 0.9 - padding[0];
-			if ( nullptr != gPad) y2 = 1.0 - gPad->GetTopMargin() - padding[0];
-			y1 = y2 - h;
-		}
-		if ( spos.find("right") != string::npos ){
-			LOG_F( INFO, "LEGEND: RIGHT" );
-			x2 = 0.9 - padding[1];
-			if ( nullptr != gPad) x2 = 1.0 - gPad->GetRightMargin() - padding[1];
-			x1 = x2 - w;
-		}
-		if ( spos.find("bottom") != string::npos ){
-			LOG_F( INFO, "LEGEND: BOTTOM" );
-			y1 = 0.1 + padding[2];
-			if ( nullptr != gPad) y1 = gPad->GetBottomMargin() + padding[2];
-			y2 = y1 + h;
-		}
-		if ( spos.find("left") != string::npos ){
-			LOG_F( INFO, "LEGEND: LEFT" );
-			x1 = 0.1 + padding[3];
-			if ( nullptr != gPad) x1 = gPad->GetLeftMargin() + padding[3];
-			x2 = x1 + w;
-		}
-		if ( spos.find("vcenter") != string::npos ){
-			LOG_F( INFO, "LEGEND: VERTICAL CENTER" );
-			y1 = 0.5 - h/2;
-			y2 = 0.5 + h/2;
-		}
-		if ( spos.find("hcenter") != string::npos ){
-			LOG_F( INFO, "LEGEND: HORIZONTAL CENTER" );
-			x1 = 0.5 - w/2;
-			x2 = 0.5 + w/2;
-		}
-		DLOG( "Legend position: (%f, %f) -> (%f, %f)", x1, y1, x2, y2 );
-		
-		
-
-		TLegend * leg = new TLegend( x1, y1, x2, y2 );
-
-		if ( config.exists(  _path + ".Legend:title" ) ) {
-			leg->SetHeader( config.getXString( _path + ".Legend:title" ).c_str() );
-		}
-		leg->SetNColumns( config.getInt( _path + ".Legend:columns", 1 ) );
-
-		vector<string> entries = config.childrenOf( _path + ".Legend", "Entry" );
-
-
-		// HISTOGRAMS
-		for ( string entryPath : entries ){
-			//INFO( classname(), "Entry @" << entryPath );
-			if ( config.exists( entryPath + ":name" ) != true ) continue;
-			string name = config.getXString( entryPath + ":name" );
-			
-			if ( histos.count( name ) <= 0 || histos[ name ] == nullptr ) {
-				continue;
-			}
-			TH1 * h = (TH1*)histos[ name ]->Clone( ("hist_legend_" + name).c_str() );
-			//INFO( classname(), "Entry histo=" << h );
-			string t = config.getXString(  entryPath + ":title", name );
-			string opt = config.getXString(  entryPath + ":opt", "l" );
-
-			// if ( config.exists( entryPath + ":markersize" ) ){
-			rpl.style( h ).set( config, entryPath );
-			// }
-
-
-			leg->AddEntry( h, t.c_str(), opt.c_str() );
-		}
-
-		// GRAPHS
-		for ( string entryPath : entries ){
-			//INFO( classname(), "Entry @" << entryPath );
-			if ( config.exists( entryPath + ":name" ) != true ) continue;
-			string name = config.getXString( entryPath + ":name" );
-			
-			if ( graphs.count( name ) <= 0 || graphs[ name ] == nullptr ) {
-				continue;
-			}
-			TGraph * g = (TGraph*)graphs[ name ]->Clone( ("graph_legend_" + name).c_str() );
-
-
-			//INFO( classname(), "Entry histo=" << h );
-			string t = config.getXString(  entryPath + ":title", name );
-			string opt = config.getXString(  entryPath + ":opt", "l" );
-
-			rpl.style( g ).set( config, entryPath );
-			leg->AddEntry( g, t.c_str(), opt.c_str() );
-		}
-
-		// FUNCTIONS
-		for ( string entryPath : entries ){
-			//INFO( classname(), "Entry @" << entryPath );
-			if ( config.exists( entryPath + ":name" ) != true ) continue;
-			string name = config.getXString( entryPath + ":name" );
-			
-			if ( funcs.count( name ) <= 0 || funcs[ name ] == nullptr ) {
-				continue;
-			}
-			TF1 * f = (TF1*)funcs[ name ]->Clone( ("graph_legend_" + name).c_str() );
-
-
-			//INFO( classname(), "Entry histo=" << h );
-			string t = config.getXString(  entryPath + ":title", name );
-			string opt = config.getXString(  entryPath + ":opt", "l" );
-
-			rpl.style( f ).set( config, entryPath );
-			leg->AddEntry( f, t.c_str(), opt.c_str() );
-		}
-
-		if ( entries.size() <= 0 ){
-			// Add an entry for all histograms by default if no <Entry> nodes found
-			for ( auto kv : histos ){
-				string t = config.getString(  _path + ".Legend." + kv.first+ ":title", kv.first) ;
-				string opt = config.getString(  _path + ".Legend." + kv.first+ ":opt", "l") ;
-				leg->AddEntry( kv.second, t.c_str(), opt.c_str() );
-			}
-		}
-
-		if ( config.exists( _path + ".Legend:text_size" ) ){
-			leg->SetTextSize( config.getDouble( _path + ".Legend:text_size" ) );
-		}
-		if ( config.exists( _path + ".Legend:text_point" ) ){
-			leg->SetTextSize( config.getDouble( _path + ".Legend:text_point" ) / 360.0 );
-		}
-		if ( config.exists( _path + ".Legend:border_size" ) ){
-			leg->SetBorderSize( config.getDouble( _path + ".Legend:border_size" ) );
-		}
-		if ( config.exists( _path + ".Legend:fill_color" ) ){
-			leg->SetFillColor( color( config.getString( _path + ".Legend:fill_color" ) ) );
-		}
-		if ( config.exists( _path + ".Legend:fill_style" ) ){
-			//INFOC( "FillStyle=" << config.getInt( _path + ".Legend:fill_style" ) );
-			leg->SetFillStyle( config.getInt( _path + ".Legend:fill_style" ) );
-		}
-
-		leg->Draw( );
-
-	}
-}
-
-void VegaXmlPlotter::makeLatex( string _path ){
-	DSCOPE();
-	//INFO( classname(), "_path=" << quote(_path) );
-	
-	TLatex latex;
-	vector<string> latex_paths = config.childrenOf( _path, "TLatex", 1 );
-	for ( string ltpath : latex_paths ){
-		if ( !config.exists( ltpath + ":text" ) ) continue;
-
-		string text = config.getXString( ltpath + ":text" );
-		
-		//INFO( classname(), "Latex @ " << ltpath );
-		latex.SetTextSize( config.getFloat( ltpath + ":size", 0.05 ) );
-		if ( config.exists( ltpath + ":point" ) )
-			latex.SetTextSize( config.getFloat( ltpath + ":point", 18 ) / 360.0 );
-		
-
-
-		if ( config.exists( ltpath +":color" ) ){
-			string cs = config.getXString( ltpath + ":color" );
-			latex.SetTextColor( color( cs ) );
-		} else {
-			latex.SetTextColor( kBlack );
-		}
-		
-		TLatex * TL = nullptr;
-		if ( config.exists( ltpath + ":x" ) && config.exists( ltpath + ":y" ) ){
-			TL = latex.DrawLatexNDC( config.getFloat( ltpath + ":x" ), 
-							config.getFloat( ltpath + ":y" ), 
-							config.getXString( ltpath + ":text" ).c_str() );
-		} else if ( config.exists( ltpath + ":ux" ) && config.exists( ltpath + ":uy" ) ){
-			TL = latex.DrawLatex( config.getFloat( ltpath + ":ux" ), 
-							config.getFloat( ltpath + ":uy" ), 
-							config.getXString( ltpath + ":text" ).c_str() );
-		}
-		if ( nullptr == TL ){
-			LOG_F(ERROR, "Cannot draw latex, missing attribute. Must have pad corrds (:x,:y) or user corrds( :ux, :uy)" ); 
-		}		 
-
-		if ( config.exists( ltpath + ":font" ) ){
-			TL->SetTextFont( config.getInt( ltpath + ":font" ) );
-		}
-		if ( config.exists( ltpath + ":angle" ) ){
-			TL->SetTextAngle( config.getFloat( ltpath + ":angle" ) );
-		}
-
-		if ( config.exists( ltpath + ":align" ) ){
-			string strAlign = config.getString( ltpath + ":align" );
-			if ( "top" == strAlign || "t" == strAlign ){
-				TL->SetTextAlign( 13 );
-			} else if ( "center" == strAlign || "centered" == strAlign || "c" == strAlign ){
-				TL->SetTextAlign( 12 );
-			} else if ( "bottom" == strAlign || "b" == strAlign ){
-				TL->SetTextAlign( 11 );
-			} else if ( "special" == strAlign || "s" == strAlign ){
-				TL->SetTextAlign( 10 );
-			} else {
-				TL->SetTextAlign( config.getInt( ltpath + ":align" ) );
-			}
-		}
-
-	}
-}
-
-void VegaXmlPlotter::makeLine( string _path ){
-	DSCOPE();
-	//INFO( classname(), "_path=" << quote(_path) );
-
-	vector<string> line_paths = config.childrenOf( _path, "TLine", 1 );
-	for ( string ltpath : line_paths ){
-		
-		//INFO( classname(), "Line (" << config.getDouble( ltpath + ":x1", 0.0 ) << ", " << config.getDouble( ltpath + ":y1", 0.0 ) << ") -> (" << config.getDouble( ltpath + ":x2", 0.0 ) << ", " << config.getDouble( ltpath + ":y2", 0.0 ) << ")" )
-
-		vector<double> x(2);
-		vector<double> y(2);
-
-		if ( config.exists( ltpath + ":x" ) ){
-			x = config.getDoubleVector( ltpath + ":x" );
-		}
-		if ( config.exists( ltpath + ":y" ) ){
-			y = config.getDoubleVector( ltpath + ":y" );
-		}
-		if ( config.exists( ltpath + ":p1" ) ){
-			vector<double> p = config.getDoubleVector( ltpath + ":p1" );
-			if ( p.size() >= 2 ){
-				x[0] = p[0];
-				y[0] = p[1];
-			}
-		}
-		if ( config.exists( ltpath + ":p2" ) ){
-			vector<double> p = config.getDoubleVector( ltpath + ":p2" );
-			if ( p.size() >= 2 ){
-				x[1] = p[0];
-				y[1] = p[1];
-			}
-		}
-
-		LOG_F( INFO, "Line (%f, %f)->(%f, %f)", x[0], y[0], x[1], y[1] );
-		TLine * line = new TLine( 
-			config.getDouble( ltpath + ":x1", x[0] ),
-			config.getDouble( ltpath + ":y1", y[0] ),
-			config.getDouble( ltpath + ":x2", x[1] ),
-			config.getDouble( ltpath + ":y2", y[1] ) );
-
-		//INFO( classname(), "TLine @ " << ltpath );
-		
-		// TODO: add string color support here
-		line->SetLineColor( color( config.getString( ltpath + ":color" ) ) );
-		line->SetLineWidth( config.getInt( ltpath + ":width", 1 ) );
-		line->SetLineStyle( config.getInt( ltpath + ":style", 2 ) );
-
-		line->Draw("same");
-	}
-
-}
-
-void VegaXmlPlotter::makeExports( string _path, TPad * _pad ){
-	DSCOPE();
-	vector<string> exp_paths = config.childrenOf( _path, "Export", 1 );
-	if ( nullptr == _pad ) {
-		_pad = (TPad*)gPad;
-	}
-	if ( nullptr == _pad ) return; 
-	for ( string epath : exp_paths ){
-		if ( !config.exists( epath + ":url" ) ) continue;
-		string url = config.getXString( epath + ":url" );
-		if ( url.find( ".json" ) != string::npos ){
-			// TBufferJSON::ExportToFile( url.c_str(), _pad );
-			// LOG_F( INFO, "%s",  );
-			// ofstream fout( url.c_str() );
-			// fout << TBufferJSON::ConvertToJSON( _pad ).Data();
-			// fout.close();
-		} else {
-			_pad->Print( url.c_str() );
-		}
-		
-	}
-}
+} // makeHistogram
 
 TH1* VegaXmlPlotter::makeHistoFromDataTree( string _path, int iHist ){
 	DSCOPE();
@@ -1685,7 +493,7 @@ TH1* VegaXmlPlotter::makeHistoFromDataTree( string _path, int iHist ){
 	globalHistos[ hName ] = h;
 
 	return h;
-}
+} // makeHistoFromDataTree
 
 void VegaXmlPlotter::positionOptStats( string _path, TPaveStats * st ){
 	DSCOPE();
@@ -1712,730 +520,7 @@ void VegaXmlPlotter::positionOptStats( string _path, TPaveStats * st ){
 		st->SetY1NDC( config.getFloat( _path + ".OptStats:y1" ) );
 	if ( config.exists( _path + ".OptStats:y2" ) )
 		st->SetY2NDC( config.getFloat( _path + ".OptStats:y2" ) ); 
-}
-
-TCanvas* VegaXmlPlotter::makeCanvas( string _path ){
-	DSCOPE();
-	int width = -1, height = -1;
-	// global first
-	width = config.getInt( "TCanvas:width", -1 );
-	height = config.getInt( "TCanvas:height", -1 );
-
-	width = config.getInt( _path + ".TCanvas:width", width );
-	height = config.getInt( _path + ".TCanvas:height", height );
-
-	TCanvas * c = nullptr; 
-	if ( width != -1 && height > -1 )
-		c = new TCanvas( "c", "c", width, height );
-	else 
-		c = new TCanvas( "c" );
-
-	c->SetFillColor(0); 
-	c->SetFillStyle(4000);
-
-	return c;
-}
-
-
-void VegaXmlPlotter::makeTransforms(  ){
-	DSCOPE();
-	vector<string> tform_paths = config.childrenOf( nodePath, "Transform" );
-	LOG_F( INFO, "Found %lu Transforms", tform_paths.size() );
-	//INFOC( "Found " << tform_paths.size() << plural( tform_paths.size(), " Transform set", " Transform sets" ) );
-	
-
-	for ( string path : tform_paths ){
-		vector<string> states = config.getStringVector( path + ":states" );
-		int i = 0;
-		if( states.size() == 0 ) states.push_back( "" );
-
-
-		for ( string state : states ){
-			// set the global vars for this template
-			config.set( "state", state );
-			config.set( "i", ts(i) );
-			makeTransform( path );
-			i++;
-		}
-	}
-}
-
-void VegaXmlPlotter::makeTransform( string tpath ){
-	DSCOPE();
-	vector<string> tform_paths = config.childrenOf( tpath );
-	LOG_F( INFO, "Found %lu Transform paths", tform_paths.size() );
-	//INFOC( "Found " << tform_paths.size() << plural( tform_paths.size(), " Transform", " Transforms" ) );
-
-	// TODO function map from string to transform??
-	for ( string tform : tform_paths ){
-		string tn = config.tagName( tform );
-		//:( tn );
-
-		if ( "Projection" == tn )
-			makeProjection( tform );
-		if ( "ProjectionX" == tn )
-			makeProjectionX( tform );
-		if ( "ProjectionY" == tn )
-			makeProjectionY( tform );
-		if ( "Add" == tn )
-			makeAdd( tform );
-		if ( "Divide" == tn )
-			makeDivide( tform );
-		if ( "Rebin" == tn )
-			makeRebin( tform );
-		if ( "Scale" == tn )
-			makeScale( tform );
-		if ( "Clone" == tn )
-			makeClone( tform );
-		if ( "Draw" == tn )
-			makeDraw( tform );
-		if ( "Smooth" == tn )
-			makeSmooth( tform );
-		if ( "CDF" == tn )
-			makeCDF( tform );
-		if ( "Style" == tn )
-			transformStyle( tform );
-		if ( "SetBinError" == tn )
-			makeSetBinError( tform );
-		if ( "BinLabels" == tn || "SetBinLabels" == tn )
-			makeBinLabels( tform );
-		
-
-	}
-}
-
-void VegaXmlPlotter::makeSetBinError( string _path ){
-	DSCOPE();
-	
-	TH1 * h = findHistogram( _path, 0 );
-	TH2 *h2 = dynamic_cast<TH2*>(h);
-	TH3 *h3 = dynamic_cast<TH3*>(h);
-
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "Histogram is not valid for Transform @ %s", _path.c_str() );
-		return;
-	}
-
-	float value = config.getFloat( _path + ":value", 0 );
-
-	if ( nullptr != h3  ){
-		for ( int i = 1; i <= h3->GetXaxis()->GetNbins(); i++ ){
-			for ( int j = 1; j <= h3->GetYaxis()->GetNbins(); j++ ){
-				for ( int k = 1; k <= h3->GetZaxis()->GetNbins(); k++ ){
-					int b = h3->GetBin( i, j, k );
-					h3->SetBinError( b, value );
-				}
-			}
-		}	
-	}
-	if ( nullptr != h2  ){
-		for ( int i = 1; i <= h2->GetXaxis()->GetNbins(); i++ ){
-			for ( int j = 1; j <= h2->GetYaxis()->GetNbins(); j++ ){
-				int b = h2->GetBin( i, j );
-				h2->SetBinError( b, value );
-			}
-		}	
-	}
-	for ( int i = 1; i <= h->GetXaxis()->GetNbins(); i++ ){
-		h->SetBinError( i, value );
-	}
-}
-
-
-void VegaXmlPlotter::makeProjection( string _path ){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	string nn = config.getXString( _path + ":save_as" );
-	string axis = config.getXString( _path + ":axis", "x" );
-
-	TH1 * h = findHistogram( _path, 0 );
-	TH2 *h2 = dynamic_cast<TH2*>(h);
-	TH3 *h3 = dynamic_cast<TH3*>(h);
-	
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "Histogram is not valid for Transform @ %s", _path.c_str() );
-		return;
-	}
-
-	if ( nullptr != h3 ){
-		LOG_F( INFO, "Projecting TH3" );
-		if ( "x" == axis || "X" == axis ){
-			LOG_F( INFO, "Projecting 1D onto %s Axis", axis.c_str() );
-			int by1 = getProjectionBin( _path, h, "y", "1",  0 );
-			int by2 = getProjectionBin( _path, h, "y", "2", -1 );
-
-			int bz1 = getProjectionBin( _path, h, "z", "1",  0 );
-			int bz2 = getProjectionBin( _path, h, "z", "2", -1 );
-
-			TH1 * hNew = h3->ProjectionX( nn.c_str(), by1, by2, bz1, bz2 );
-			globalHistos[ nn ] = hNew;
-		} else if ( "y" == axis || "Y" == axis ){
-			LOG_F( INFO, "Projecting 1D onto %s Axis", axis.c_str() );
-			int bx1 = getProjectionBin( _path, h, "x", "1",  0 );
-			int bx2 = getProjectionBin( _path, h, "x", "2", -1 );
-
-			int bz1 = getProjectionBin( _path, h, "z", "1",  0 );
-			int bz2 = getProjectionBin( _path, h, "z", "2", -1 );
-
-			TH1 * hNew = h3->ProjectionY( nn.c_str(), bx1, bx2, bz1, bz2 );
-			globalHistos[ nn ] = hNew;
-		} else if ( "z" == axis || "Z" == axis ){
-			LOG_F( INFO, "Projecting 1D onto %s Axis", axis.c_str() );
-			int bx1 = getProjectionBin( _path, h, "x", "1",  0 );
-			int bx2 = getProjectionBin( _path, h, "x", "2", -1 );
-
-			int by1 = getProjectionBin( _path, h, "y", "1",  0 );
-			int by2 = getProjectionBin( _path, h, "y", "2", -1 );
-			LOG_F( INFO, "ProjectionZ : x(%d, %d), y : (%d, %d)", bx1, bx2, by1, by2 );
-			TH1 * hNew = h3->ProjectionZ( nn.c_str(), bx1, bx2, by1, by2 );
-			globalHistos[ nn ] = hNew;
-		} else {
-			// lets do a projection in 3D
-			int bx1 = getProjectionBin( _path, h, "x", "1",  0 );
-			int bx2 = getProjectionBin( _path, h, "x", "2", -1 );
-
-			int by1 = getProjectionBin( _path, h, "y", "1",  0 );
-			int by2 = getProjectionBin( _path, h, "y", "2", -1 );
-
-			int bz1 = getProjectionBin( _path, h, "z", "1",  0 );
-			int bz2 = getProjectionBin( _path, h, "z", "2", -1 );
-
-			if ( axesRangeGiven( _path, "x" ) ){
-				LOG_F( INFO, "x : (%d, %d)", bx1, bx2 );
-				h3->GetXaxis()->SetRange( bx1, bx2 );
-			} else 
-				h3->GetXaxis()->SetRange( 1, -1 );
-			if ( axesRangeGiven( _path, "y" ) ){
-				LOG_F( INFO, "y : (%d, %d)", by1, by2 );
-				h3->GetYaxis()->SetRange( by1, by2 );
-			} else 
-				h3->GetYaxis()->SetRange( 1, -1 );
-			if ( axesRangeGiven( _path, "z" ) ){
-				LOG_F( INFO, "z : (%d, %d)", bz1, bz2 );
-				h3->GetZaxis()->SetRange( bz1, bz2 );
-			} else 
-				h3->GetZaxis()->SetRange( 1, -1 );
-			
-			TH1 * hNew = (TH1*)h3->Project3D( axis.c_str() );
-			hNew->SetName( nn.c_str() );
-			globalHistos[ nn ] = hNew;
-		} // else 2D projection
-
-	} else if ( nullptr != h2 ){
-		if ( "x" == axis || "X" == axis )
-			return makeProjectionX( _path );
-		else 
-			return makeProjectionY( _path );
-	}
-
-}
-
-
-void VegaXmlPlotter::makeProjectionX( string _path){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	string nn = config.getXString( _path + ":save_as" );
-	
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		// ERRORC( "Could not find histogram " << quote( d + "/" + n ) );
-		return;
-	}
-
-	double y1 = -999;
-	int b1 = config.getInt( _path + ":b1", 0 );
-	if ( config.exists( _path + ":y1" ) ){
-		y1 = config.getDouble( _path + ":y1", -1 );
-		//INFOC( "ProjectionX y1=" << y1 );
-		b1 = ((TH2*)h)->GetYaxis()->FindBin( y1 );
-	}
-	
-	double y2 = -999;
-	int b2 = config.getInt( _path + ":b2", -1 );
-	if ( config.exists( _path + ":y2" ) ){
-		y2 = config.getDouble( _path + ":y2", -1 );
-		//INFOC( "ProjectionX y2=" << y2 );
-		b2 = ((TH2*)h)->GetYaxis()->FindBin( y2 );
-	}
-	//INFOC( "ProjectionX [" << nn << "] b1=" << b1 << ", b2="<<b2 );
-
-	double step = config.getDouble( _path + ":step", -1.0 );
-	if ( step > 0.0 && y1 != -999 && y2 != -999 ){
-		for (double y = y1; y <= y2; y += step ){
-			config.set( _path + ":step", "" ); // prevent inf recursion
-			config.set( _path +":save_as", nn + "_" + dtes( y, "p" ) + "_" + dtes( y + step, "p" ) );
-			config.set( _path + ":y1", dts( y ) );
-			config.set( _path + ":y2", dts( y + step ) );
-
-			makeProjectionX( _path );
-		}
-		return;
-	} 
-
-	TH1 * hOther = ((TH2*)h)->ProjectionX( nn.c_str(), b1, b2 );
-	h = hOther;
-
-	globalHistos[ nn ] = h;
-}
-
-
-
-void VegaXmlPlotter::makeProjectionY( string _path){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		// ERRORC( "Could not find histogram " << quote( d + "/" + n ) );
-		return;
-	}
-	string nn = config.getXString( _path + ":save_as" );
-
-	int b1 = config.getInt( _path + ":b1", 0 );
-	if ( config.exists( _path + ":x1" ) ){
-		double x1 = config.getDouble( _path + ":x1", -1 );
-		// INFOC( "ProjectionY x1=" << x1 );
-		b1 = ((TH2*)h)->GetYaxis()->FindBin( x1 );
-	}
-	
-	int b2 = config.getInt( _path + ":b2", -1 );
-	if ( config.exists( _path + ":x2" ) ){
-		double x2 = config.getDouble( _path + ":x2", -1 );
-		// INFOC( "ProjectionY x2=" << x2 );
-		b2 = ((TH2*)h)->GetYaxis()->FindBin( x2 );
-	}
-	//INFOC( "ProjectionX [" << nn << "] b1=" << b1 << ", b2="<<b2 );
-
-	TH1 * hOther = ((TH2*)h)->ProjectionY( nn.c_str(), b1, b2 );
-	h = hOther;
-
-	globalHistos[ nn ] = h;
-
-}
-
-void VegaXmlPlotter::makeMultiAdd( string _path ){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-
-	string d = config.getXString( _path + ":data" );
-	vector<string> n = config.getStringVector( _path + ":name" );
-	if ( n.size() < 1 ) {
-		n = config.getStringVector( _path + ":names" );
-	}
-
-	string nn = config.getXString( _path + ":save_as" );
-
-	// TODO allow modifier for each
-	// double mod = config.getDouble( _path + ":mod", 1.0 );
-
-	TH1 * hFirst = findHistogram( d, n[0] );
-	if ( nullptr == hFirst ){
-		// ERRORC( "Hit Null Histo " << n[0] );
-		return;
-	}
-	TH1 * hSum = (TH1*)hFirst->Clone( nn.c_str() );
-
-	for ( int i = 1; i < n.size(); i++ ){
-		TH1 * h = findHistogram( d, n[i] );
-		if ( nullptr == h ) {
-			LOG_F( WARNING, "Cannot add n=(%s), nullptr", n[i].c_str() );
-			continue;
-		}
-		hSum ->Add( h );
-	}
-
-	globalHistos[nn] = hSum;
-}
-
-void VegaXmlPlotter::makeAdd( string _path){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	if ( config.exists( _path+":name" ) && config.getStringVector( _path+":name" ).size() > 1 ){
-		makeMultiAdd( _path );
-		return;
-	}
-	if ( config.exists( _path+":names" ) && config.getStringVector( _path+":names" ).size() > 1 ){
-		makeMultiAdd( _path );
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	string nn = config.getXString( _path + ":save_as" );
-	double mod = config.getDouble( _path + ":mod", 1.0 );
-
-	TH1 * hA = findHistogram( _path, 0, "A" );
-	TH1 * hB = findHistogram( _path, 0, "B" );
-	if ( nullptr == hA || nullptr == hB ) {
-		// ERRORC( "Could not find histogram" );
-		return;
-	}
-
-	TH1 * hSum = (TH1*) hA->Clone( nn.c_str() );
-	hSum->Add( hB, mod );
-	globalHistos[nn] = hSum;
-}
-void VegaXmlPlotter::makeDivide( string _path){
-	DSCOPE();
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	string nn = config.getXString( _path + ":save_as" );
-
-	TH1 * hNum = findHistogram( _path, 0, "A" );
-	TH1 * hDen = findHistogram( _path, 0, "B" );
-	if ( nullptr == hNum || nullptr == hDen ) {
-		// ERRORC( "Could not find histogram" );
-		return;
-	}
-
-	TH1 * hOther = (TH1*) hNum->Clone( nn.c_str() );
-	hOther->Divide( hDen );
-	globalHistos[nn] = hOther;
-}
-
-void VegaXmlPlotter::makeRebin( string _path ){
-	DSCOPE();
-	
-	// rebins to array of bin edges
-	bool in_place = false;
-	if ( !config.exists( _path + ":save_as" ) ){
-		LOG_F( WARNING, "Rebin does not support in-place!" );
-		in_place = true;
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		// ERRORC( "Could not find histogram " << quote( d + "/" + n ) );
-		LOG_F( ERROR, "Could not find histogram " );
-		return;
-	}
-
-
-	string nn = config.getXString( _path + ":save_as" );
-	TH1 * hOther = nullptr;//((TH1*)h)->Clone( nn.c_str() );
-
-	HistoBins bx, by, bz;
-	if ( config.exists( _path +":bins_x" )  ){
-		if ( config.exists( config.getXString( _path+":bins_x" ) ) ){
-			bx.load( config, config.getXString( _path+":bins_x" ) );
-		} else {
-			bx.load( config, _path + ":bins_x" );
-		}
-	}
-	if ( config.exists( _path +":bins_y" )  ){
-		if ( config.exists( config.getXString( _path+":bins_y" ) ) ){
-			by.load( config, config.getXString( _path+":bins_y" ) );
-		} else {
-			by.load( config, _path + ":bins_y" );
-		}
-	}
-	if ( config.exists( _path +":bins_z" )  ){
-		if ( config.exists( config.getXString( _path+":bins_z" ) ) ){
-			bz.load( config, config.getXString( _path+":bins_z" ) );
-		} else {
-			bz.load( config, _path + ":bins_z" );
-		}
-	}
-
-	int nDim = 0;
-	if ( bx.nBins() > 0 ) nDim++;
-	if ( by.nBins() > 0 ) nDim++;
-	if ( bz.nBins() > 0 ) nDim++;
-
-	DLOG( "nDim=%d", nDim );
-
-
-	if ( nDim == 1 && bx.nBins() > 0 ){
-		hOther = h->Rebin( bx.nBins(), nn.c_str(), bx.bins.data() );
-		globalHistos[nn] = hOther;
-	} else if ( nDim == 1 ){
-		// ERRORC( "Cannot Rebin, check error message for x bins" );
-	}
-
-	if ( nDim == 2 && bx.nBins() > 0 && by.nBins() > 0 && nullptr != dynamic_cast<TH2*>( h ) ){
-		hOther = new TH2D( nn.c_str(), h->GetTitle(), bx.nBins(), bx.bins.data(), by.nBins(), by.bins.data() );
-		HistoBins::rebin2D( dynamic_cast<TH2*>(h), dynamic_cast<TH2*>(hOther) );
-		globalHistos[nn] = hOther;
-	} else if ( nDim == 2 ){
-		if ( nullptr == dynamic_cast<TH2*>( h ) ){
-			LOG_F( ERROR, "Input histogram is not 2D" );
-		} else {
-			LOG_F( ERROR, "Cannot Rebin, check error for bins x and y" );
-		}
-	}
-
-	if ( nDim == 3 && bx.nBins() > 0 && by.nBins() > 0 && bz.nBins() > 0 && nullptr != dynamic_cast<TH3*>( h ) ){
-		hOther = new TH3D( nn.c_str(), h->GetTitle(), bx.nBins(), bx.bins.data(), by.nBins(), by.bins.data(), bz.nBins(), bz.bins.data() );
-		HistoBins::rebin3D( dynamic_cast<TH3*>(h), dynamic_cast<TH3*>(hOther) );
-		globalHistos[nn] = hOther;
-	} else if ( nDim == 3 ){
-		if ( nullptr == dynamic_cast<TH2*>( h ) ){
-			LOG_F( ERROR, "Input histogram is not 3D" );
-		} else {
-			LOG_F( ERROR, "Cannot Rebin, check error for bins x, y, z" );
-		}
-	}
-}
-
-void VegaXmlPlotter::makeScale( string _path ){
-	DSCOPE();
-
-	if ( !config.exists( _path + ":save_as" ) ){
-		// ERRORC( "Must provide " << quote( "save_as" ) << " attribute to save transformation" );
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		// ERRORC( "Could not find histogram " << quote( d + "/" + n ) );
-		return;
-	}
-	string nn = config.getXString( _path + ":save_as" );
-	TH1 * hOther = (TH1*)h->Clone( nn.c_str() );
-
-	double factor = config.getDouble( _path + ":factor", 1.0 );
-	string opt    = config.getXString( _path +":opt", "" );
-
-	hOther->Scale( factor, opt.c_str() );
-	globalHistos[nn] = hOther;
-}
-
-void VegaXmlPlotter::makeDraw( string _path ){
-	DSCOPE();
-
-	string d = config.getXString( _path + ":data" );
-	string nn = nameOnly( config.getXString( _path + ":name" ) );
-	TH1 * h = makeHistoFromDataTree( _path, 0 ); //findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "Could not make %s", nn.c_str() );
-		return;
-	}
-
-	// string nn = config.getXString( _path + ":save_as" );
-	globalHistos[ nn ] = h;
-}
-
-
-void VegaXmlPlotter::makeSmooth( string _path ){
-	DSCOPE();
-
-	bool in_place = false;
-	if ( !config.exists( _path + ":save_as" ) ){
-		LOG_F( WARNING, "Smoothing histogram in place!" );
-		in_place = true;
-	}
-
-	string d = config.getString( _path + ":data" );
-	string n = config.getString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "could not find histo %s %s", d.c_str(), n.c_str() );
-		return;
-	}
-
-	int nSmooth = config.get<int>( _path + ":n", 1 );
-	
-	if ( in_place ){
-		LOG_F( INFO, "Smoothing histogram %d times", nSmooth );
-		h->Smooth( nSmooth );
-	} else {
-		string nn = config.getString( _path + ":save_as" );
-		TH1 * hOther = (TH1*)h->Clone( nn.c_str() );
-		hOther->Smooth( nSmooth );
-		LOG_F( INFO, "Smoothing histogram %s %d times", nn.c_str(), nSmooth );
-		globalHistos[nn] = hOther;
-	}
-}
-
-
-void VegaXmlPlotter::makeCDF( string _path ){
-	DSCOPE();
-
-	
-	if ( !config.exists( _path + ":save_as" ) ){
-		LOG_F( WARNING, "Cannot make CDF in place, abort" );
-		return;
-	}
-
-	string d = config.getString( _path + ":data" );
-	string n = config.getString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "could not find histo %s %s", d.c_str(), n.c_str() );
-		return;
-	}
-
-	bool forward = config.get<bool>( _path + ":forward", false );
-	
-	string nn = config.getString( _path + ":save_as" );
-	TH1 * hOther = (TH1*)h->Clone( nn.c_str() );
-	hOther = hOther->GetCumulative( forward, "" );
-	LOG_F( INFO, "Made CDF for histogram %s, with forward=%s", nn.c_str(), bts(forward).c_str() );
-	globalHistos[nn] = hOther;
-	
-}
-
-
-void VegaXmlPlotter::makeBinLabels( string _path ){
-	LOG_F( INFO, "" );
-	if ( !config.exists( _path + ":save_as" ) ){
-		LOG_F( WARNING, "Cannot make CDF in place, abort" );
-		return;
-	}
-
-	string d = config.getString( _path + ":data" );
-	string n = config.getString( _path + ":name" );
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		LOG_F( ERROR, "could not find histo %s %s", d.c_str(), n.c_str() );
-		return;
-	}
-	
-	string nn = config.getString( _path + ":save_as" );
-	TH1 * hOther = (TH1*)h->Clone( nn.c_str() );
-
-	vector<string> labels = config.getStringVector( _path + ":labels" );
-	TAxis *axis = hOther->GetXaxis();
-
-	for ( size_t i = 0; i < labels.size(); i++ ){
-		axis->SetBinLabel( i+1, labels[i].c_str() );
-		LOG_F( INFO, "Seeting bin[%lu] label to \"%s\" ", i+1, labels[i].c_str() );
-	}
-	// LOG_F( INFO, "Made CDF for histogram %s, with forward=%s", nn.c_str(), bts(forward).c_str() );
-	globalHistos[nn] = hOther;
-}
-
-
-
-void VegaXmlPlotter::makeClone( string _path ){
-
-	if ( !config.exists( _path + ":save_as" ) ){
-		return;
-	}
-
-	string d = config.getXString( _path + ":data" );
-	string n = config.getXString( _path + ":name" );
-
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		return;
-	}
-
-	config.set( "uname", underscape(n) );
-
-	string nn = config.getXString( _path + ":save_as" );
-	TH1 * hOther = (TH1*)h->Clone( nn.c_str() );
-
-	// if bin ranges or axis ranges exist then clear and clone subrange
-	if ( config.exists( _path+":b1" ) || config.exists( _path+":x1" ) ){
-		hOther->Reset();
-		int b1 = config.getInt( _path + ":b1", 0 );
-		if ( config.exists( _path + ":x1" ) ){
-			double x1 = config.getDouble( _path + ":x1", -1 );
-			b1 = h->GetXaxis()->FindBin( x1 );
-		}
-		
-		int b2 = config.getInt( _path + ":b2", -1 );
-		if ( config.exists( _path + ":x2" ) ){
-			double x2 = config.getDouble( _path + ":x2", -1 );
-			b2 = h->GetXaxis()->FindBin( x2 );
-		}
-
-		// now clone the subrange
-		HistoBook::cloneBinRange( h, hOther, b1, b2 );
-	}
-
-	globalHistos[nn] = hOther;
-}
-
-void VegaXmlPlotter::transformStyle( string _path ){
-	DSCOPE();
-	TH1 * h = findHistogram( _path, 0 );
-	if ( nullptr == h ) {
-		return;
-	}
-
-	RooPlotLib rpl;
-
-	DLOG( "settign style from: %s, and %s", (_path + ":style").c_str(), (_path + ".style").c_str() );
-	rpl.style( h ).set( config, _path).set( config, _path + ":style" ).set( config, _path + ".style" );
-
-}
-
-
-
-// void VegaXmlPlotter::drawCanvas( string _path ){
-// 	DSCOPE();
-
-// 	XmlCanvas xcanvas( config, _path );
-	
-// 	vector<string> ppads = config.childrenOf( _path, "Pad" );
-// 	vector<string> pplots = config.childrenOf( _path, "Plots" );
-// 	vector<string> paths = ppads;
-	
-// 	if ( ppads.size() <= 0 ) paths = pplots;
-
-// 	for ( string p : paths ){
-// 		string n = config.getXString( p + ":name", "" );
-// 		//INFOC( "Painting PAD " << n );
-// 		if ( "" == n ){
-// 			// WARNC( "Skipping pad without name" );
-// 			continue;
-// 		}
-
-// 		XmlPad * xpad = xcanvas.activatePad( n );
-
-// 		// move pad to origin
-// 		xpad->moveToOrigin();
-// 		xpad->getRootPad()->Update();
-		
-// 		makePlot( p );
-
-// 		xpad->reposition();
-// 		xpad->getRootPad()->Update();
-// 		// pad->Update();
-// 		// pad->Print( ("pad_" + n + ".pdf").c_str() );
-// 		// makeExports( p );
-
-// 	} // loop over pads
-// 	xcanvas.getCanvas()->Update();
-
-// 	makeExports( _path, xcanvas.getCanvas() );
-// }
-
+} // positionOptStats
 
 map<string, TObject*> VegaXmlPlotter::dirMap( TDirectory *dir, string prefix, bool dive ) {
 	DSCOPE();
